Remove RSP load-balancing
The actual sessions are started via Launcher, so the load on RSP is
so small, that id does not make sense to do any balancing here. If
one uses a strict active-passive LB for RSP, the internal LB of RSP
is indeed not necessary, c.f. comments by Karl in
https://rstudioide.zendesk.com/agent/tickets/49378.

The current implementation does have one major problem, though:
Sticky sessions are implemented via `ip_hash` in nginx. But since the
IP address of the RSP server stays constant, the connected Launcher
is always the same, i.e. no actual load balancing. The free version
of nginx does not support other forms of stickyness. I might try
Apache httpd in the furute.